Move the country list logic to a hook and get geo location in the same hook
