+++ client/jetpack-cloud/sections/agency-dashboard/sites-overview/dashboard-data-context.ts
@@ -9,9 +9,6 @@ const DashboardDataContext = createContext< DashboardDataContextInterface >( {
 			return undefined;
 		},
 	},
-	geoData: {
-		countryCode: '',
-	},
 } );
 
 export default DashboardDataContext;
