+++ client/jetpack-cloud/sections/agency-dashboard/sites-overview/types.ts
@@ -214,9 +214,6 @@ export interface DashboardDataContextInterface {
 		phoneNumbers: Array< string >;
 		refetchIfFailed: () => void;
 	};
-	geoData: {
-		countryCode: string;
-	};
 }
 
 export type AgencyDashboardFilterOption =
